Phase 3: Move Schemas and Create Admin
